requirements.txt
Pillow==4.0.0
python-resize-image==1.1.11
